Cleanup and fix tests for variable in test name w/ listeners
- Tests didn't work with Python 2 due to differences in string
  representation (e.g. ['foo'] vs [u'foo']).
- Test names on atest/robot side were confusing.
- Combined tests related to v2 and v3 listeners into same suite

See #2962.
