+++ atest/testdata/output/listeners/original_and_resolved_name_v2.py
@@ -1,0 +1,9 @@
+ROBOT_LISTENER_API_VERSION = 2
+
+
+def startTest(name, info):
+    print('[START] [original] %s [resolved] %s' % (info['originalname'], name))
+
+
+def end_test(name, info):
+    print('[END] [original] %s [resolved] %s' % (info['originalname'], name))
